feat(tests): enhance question bank with gift import and UI improvements
- Refactor `GiftImporter` to use a drawer component for better UX.
- Remove `QuestionsList` component and integrate its functionality into `GiftImporterDrawer`.
- Update `DetailQuestions` and `DetailQuestionsRouter` to support new question import flow.
- Adjust `ShortResponse` component to handle alternative choices more efficiently.
