+++ plugins/leemons-plugin-tests/frontend/src/components/QuestionsTablePicker/index.js
@@ -1,0 +1,1 @@
+export * from './QuestionsTablePicker';
